refactored export; requirements files
