--- requirements.txt
@@ -1,2 +1,0 @@
-requests>=2.28.0
-PyYAML>=6.0
