+++ requirements/dev-requirements.in
@@ -1,0 +1,3 @@
+-r requirements.in
+pytest
+ruff
