+++ requirements/dev-requirements.txt
@@ -1,0 +1,26 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile requirements/dev-requirements.in
+certifi==2025.6.15
+    # via requests
+charset-normalizer==3.4.2
+    # via requests
+idna==3.10
+    # via requests
+iniconfig==2.1.0
+    # via pytest
+packaging==25.0
+    # via pytest
+pluggy==1.6.0
+    # via pytest
+pygments==2.19.2
+    # via pytest
+pytest==8.4.1
+    # via -r requirements/dev-requirements.in
+pyyaml==6.0.2
+    # via -r requirements/requirements.in
+requests==2.32.4
+    # via -r requirements/requirements.in
+ruff==0.12.1
+    # via -r requirements/dev-requirements.in
+urllib3==2.5.0
+    # via requests
