+++ requirements/requirements.in
@@ -1,0 +1,2 @@
+requests
+PyYAML
