+++ requirements/requirements.txt
@@ -1,0 +1,14 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile requirements/requirements.in
+certifi==2025.6.15
+    # via requests
+charset-normalizer==3.4.2
+    # via requests
+idna==3.10
+    # via requests
+pyyaml==6.0.2
+    # via -r requirements/requirements.in
+requests==2.32.4
+    # via -r requirements/requirements.in
+urllib3==2.5.0
+    # via requests
